+++ src/Pim/Bundle/CatalogBundle/Manager/FamilyManager.php
@@ -5,7 +5,6 @@ namespace Pim\Bundle\CatalogBundle\Manager;
 use Doctrine\Common\Persistence\ObjectManager;
 use Pim\Bundle\CatalogBundle\Repository\FamilyRepositoryInterface;
 use Pim\Bundle\UserBundle\Context\UserContext;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 /**
  * Family manager
@@ -25,26 +24,21 @@ class FamilyManager
     /** @var ObjectManager */
     protected $objectManager;
 
-    /** @var EventDispatcherInterface */
-    protected $eventDispatcher;
     /**
      * Constructor
      *
      * @param FamilyRepositoryInterface $repository
      * @param UserContext               $userContext
      * @param ObjectManager             $objectManager
-     * @param EventDispatcherInterface  $eventDispatcher
      */
     public function __construct(
         FamilyRepositoryInterface $repository,
         UserContext $userContext,
-        ObjectManager $objectManager,
-        EventDispatcherInterface $eventDispatcher
+        ObjectManager $objectManager
     ) {
         $this->repository      = $repository;
         $this->userContext     = $userContext;
         $this->objectManager   = $objectManager;
-        $this->eventDispatcher = $eventDispatcher;
     }
 
     /**